fix java 1.6 compatibility, and make sure it's maintained in the future

+++ test/Reflection.java
@@ -144,8 +144,8 @@ public class Reflection {
     { Reflection r = new Reflection();
       expect(r.egads == 0);
 
-      Reflection.class.getDeclaredField("egads").set(r, 42);
-      expect(((int) Reflection.class.getDeclaredField("egads").get(r)) == 42);
+      Reflection.class.getDeclaredField("egads").set(r, (Integer)42);
+      expect(((Integer)Reflection.class.getDeclaredField("egads").get(r)) == 42);
 
       Reflection.class.getDeclaredField("egads").setInt(r, 43);
       expect(Reflection.class.getDeclaredField("egads").getInt(r) == 43);
@@ -210,7 +210,7 @@ public class Reflection {
     }
 
     try {
-      Foo.class.getField("foo").set(null, 42);
+      Foo.class.getField("foo").set(null, (Integer)42);
       expect(false);
     } catch (NoClassDefFoundError e) {
       // cool
